// packages/sdk/model/activePreset.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class ActivePreset {
  "preset"?: string;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "preset",
      baseName: "preset",
      type: "string",
    },
  ];

  static getAttributeTypeMap() {
    return ActivePreset.attributeTypeMap;
  }
}