// packages/sdk/model/addClip.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";
import { AddClipOneOf } from "./addClipOneOf";
import { AddClipOneOf1 } from "./addClipOneOf1";

export class AddClip {
  "clips"?: Array<number>;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "clips",
      baseName: "clips",
      type: "Array<number>",
    },
  ];

  static getAttributeTypeMap() {
    return AddClip.attributeTypeMap;
  }
}