// packages/sdk/model/addClipOneOf.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class AddClipOneOf {
  /**
   * Unique ID used to identify this clip
   */
  "clips"?: number;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "clips",
      baseName: "clips",
      type: "number",
    },
  ];

  static getAttributeTypeMap() {
    return AddClipOneOf.attributeTypeMap;
  }
}