// packages/sdk/model/autoExposureMode.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

/**
 * Auto exposure mode
 */
export class AutoExposureMode {
  "mode"?: AutoExposureMode.ModeEnum;
  "type"?: AutoExposureMode.TypeEnum;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "mode",
      baseName: "mode",
      type: "AutoExposureMode.ModeEnum",
    },
    {
      name: "type",
      baseName: "type",
      type: "AutoExposureMode.TypeEnum",
    },
  ];

  static getAttributeTypeMap() {
    return AutoExposureMode.attributeTypeMap;
  }
}

export namespace AutoExposureMode {
  export enum ModeEnum {
    Off = <any>"Off",
    Continuous = <any>"Continuous",
    OneShot = <any>"OneShot",
  }
  export enum TypeEnum {
    Empty = <any>"",
    Iris = <any>"Iris",
    Shutter = <any>"Shutter",
    IrisShutter = <any>"Iris,Shutter",
    ShutterIris = <any>"Shutter,Iris",
  }
}