// packages/sdk/model/codecFormat.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class CodecFormat {
  /**
   * Currently selected codec
   */
  "codec"?: string;
  /**
   * Multimedia container format
   */
  "container"?: string;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "codec",
      baseName: "codec",
      type: "string",
    },
    {
      name: "container",
      baseName: "container",
      type: "string",
    },
  ];

  static getAttributeTypeMap() {
    return CodecFormat.attributeTypeMap;
  }
}