// packages/sdk/model/contrast.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class Contrast {
  "pivot"?: number = 0.5;
  "adjust"?: number = 1.0;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "pivot",
      baseName: "pivot",
      type: "number",
    },
    {
      name: "adjust",
      baseName: "adjust",
      type: "number",
    },
  ];

  static getAttributeTypeMap() {
    return Contrast.attributeTypeMap;
  }
}