// packages/sdk/model/description.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";
import { DescriptionCapabilities } from "./descriptionCapabilities";
import { DescriptionGainRange } from "./descriptionGainRange";

export class Description {
  "gainRange"?: DescriptionGainRange;
  "capabilities"?: DescriptionCapabilities;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "gainRange",
      baseName: "gainRange",
      type: "DescriptionGainRange",
    },
    {
      name: "capabilities",
      baseName: "capabilities",
      type: "DescriptionCapabilities",
    },
  ];

  static getAttributeTypeMap() {
    return Description.attributeTypeMap;
  }
}