// packages/sdk/model/descriptionCapabilities.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";
import { DescriptionCapabilitiesPadding } from "./descriptionCapabilitiesPadding";

export class DescriptionCapabilities {
  /**
   * Input supports setting of phantom power
   */
  "phantomPower"?: boolean;
  /**
   * Input supports setting of low cut filter
   */
  "lowCutFilter"?: boolean;
  "padding"?: DescriptionCapabilitiesPadding;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "phantomPower",
      baseName: "PhantomPower",
      type: "boolean",
    },
    {
      name: "lowCutFilter",
      baseName: "LowCutFilter",
      type: "boolean",
    },
    {
      name: "padding",
      baseName: "Padding",
      type: "DescriptionCapabilitiesPadding",
    },
  ];

  static getAttributeTypeMap() {
    return DescriptionCapabilities.attributeTypeMap;
  }
}