// packages/sdk/model/descriptionCapabilitiesPadding.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class DescriptionCapabilitiesPadding {
  /**
   * Input supports setting of padding
   */
  "available"?: boolean;
  /**
   * Padding is forced to be set for the input
   */
  "forced"?: boolean;
  /**
   * Value of the padding in dB
   */
  "value"?: number;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "available",
      baseName: "available",
      type: "boolean",
    },
    {
      name: "forced",
      baseName: "forced",
      type: "boolean",
    },
    {
      name: "value",
      baseName: "value",
      type: "number",
    },
  ];

  static getAttributeTypeMap() {
    return DescriptionCapabilitiesPadding.attributeTypeMap;
  }
}