// packages/sdk/model/descriptionGainRange.ts
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class DescriptionGainRange {
  /**
   * The minimum gain value in dB
   */
  "min"?: number;
  /**
   * The maximum gain value in dB
   */
  "max"?: number;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "min",
      baseName: "Min",
      type: "number",
    },
    {
      name: "max",
      baseName: "Max",
      type: "number",
    },
  ];

  static getAttributeTypeMap() {
    return DescriptionGainRange.attributeTypeMap;
  }
}